pyproject.toml
[project]
name = "nonebot_plugin_suggarchat"
version = "3.7.0"
description = "SuggarChat chat framework"
authors = [{ name = "LiteSuggarDEV", email = "windowserror@163.com" }]
dependencies = [
    "jieba>=0.42.1",
    "nonebot2[fastapi]>=2.2.0",
    "openai>=1.66.3, <=1.99.0",
    "pydantic>=2.4.2",
    "nonebot-adapter-onebot>=2.0.0",
    "nonebot-plugin-localstore>=0.7",
    "typing-extensions>=4.6.0",
    "pytz>=2025.1",
    "tomli>=2.0.0",
    "tomli-w>=1.0.0",
    "stubs>=1.0.0",
    "dotenv>=0.9.9",
    "aiofiles>=24.1.0",
    "importlib>=1.0.4",
    "watchfiles>=0.18.0",
    "nonebot-plugin-orm>=0.8.3",
    "fastmcp>=2.13.1",
    "zipp>=3.23.0",
    "nonebot-plugin-uniconf>=0.1.1",
]
requires-python = ">=3.10, <3.14"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
keywords = ['chat", "suggar", "nonebot']

[project.optional-dependencies]
openai = []
cloudflare = ["nonebot_plugin_suggarex_cf>=2.0.0"]
full = ["nonebot_plugin_suggarex_cf>=2.0.0"]
amrita = ["amrita>=0.2.0.post2"]

[project.urls]
"Homepage" = "https://github.com/LiteSuggarDEV/nonebot_plugin_suggarchat"
"Source" = "https://github.com/LiteSuggarDEV/nonebot_plugin_suggarchat"
"Issue Tracker" = "https://github.com/LiteSuggarDEV/nonebot_plugin_suggarchat/issues"

[tool.uv]
dev-dependencies = [
    "ruff>=0.12.8",
    "nonebot-plugin-suggarex-cf>=2.0.0",
    "nonebot-plugin-orm[default]>=0.8.2",
    "pip>=25.2",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "C4",    # flake8-comprehensions
    "T10",   # flake8-debugger
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "RUF",   # Ruff-specific rules
    "I",     # isort
    "PERF",  # pylint-performance
]
ignore = [
    "E402",   # module-import-not-at-top-of-file
    "E501",   # line-too-long
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]

[tool.pyright]
typeCheckingMode = "standard"

[tool.nonebot]
plugins = ["nonebot_plugin_suggarchat", "nonebot_plugin_localstore"]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
